fix jest issues
